refactor: use CHAIN_ID
