+++ crates/constants/src/types/error.rs
