+++ crates/types/Cargo.toml
@@ -19,6 +19,7 @@ thiserror.workspace = true
 chrono.workspace = true
 
 [dev-dependencies]
+serde_json.workspace = true
 tokio = { version = "1.37.0", features = ["macros"] }
 
 [features]
